requirements.txt
httpx
apscheduler
dnspython